requirements.txt
flask==2.3.3
flask-cors==4.0.0
openai-whisper==20230314
gtts==2.3.2
pydub==0.25.1

